+++ meteortracker/camera.py
