+++ meteortracker/vid_writer.py
